+++ xroot_plugins/XrdSsiCtaRequestMessage.hpp
@@ -198,6 +198,7 @@ private:
   void processTape_Rm               (cta::xrd::Response &response);
   void processTape_Reclaim          (cta::xrd::Response &response);
   void processTape_Label            (cta::xrd::Response &response);
+  void processTapeFile_Rm           (cta::xrd::Response &response);
   void processTapePool_Add          (cta::xrd::Response &response);
   void processTapePool_Ch           (cta::xrd::Response &response);
   void processTapePool_Rm           (cta::xrd::Response &response);
@@ -207,7 +208,7 @@ private:
   void processVirtualOrganization_Add(cta::xrd::Response &response);
   void processVirtualOrganization_Ch(cta::xrd::Response &response);
   void processVirtualOrganization_Rm(cta::xrd::Response &response);
-
+  
   /*!
    * Process AdminCmd events which can return a stream response
    *