+++ ReleaseNotes.md
@@ -3,6 +3,24 @@
 ## Summary
 
 ### Features
+ - cta/CTA#1016 New options for filtering deleted files using `cta-admin rtf ls` command.
+ - cta/CTA#983 Add cta-release package for public binary rpm distribution.
+
+- [frontend] New command "tapefile rm" allows deleting a copy of a file from tape
+
+### Bug fixes
+
+ - cta/CTA#1014 Fix last column alignment when more than 1000 items are listed.
+
+# v4.0-5
+
+## Summary
+
+### Features
+- [frontend] Add options to "tapepool ls" to filter tapepools on their name, vo and encryption
+- cta/CTA#898 cta-send-event now gets the requester id and eos instance as command line arguments
+- cta/CTA#1005 "tape ls" now can filter tapes on wether they were imported from CASTOR
+- cta/CTA#1006 "repack ls" now shows the tapepool of the tape being repacked
 
 ### Bug fixes
 
@@ -58,7 +76,7 @@
 - cta/CTA#777 Minimize mounts for dual copy tape pool recalls - Temporary fix
 - cta/CTA#927 Bad message for archive route pointing non-existing pool
 - cta/CTA#930 Batched the queueing and the deleting of the repack subrequests
-- cta/CTA#969 `xrdfs query prepare` malformed JSON output 
+- cta/CTA#969 `xrdfs query prepare` malformed JSON output
 - cta/CTA#972 Updates cta-taped and rmcd man pages
 - cta/CTA#967 CI DB cleanup issues
 - cta/CTA#982 cta-fst-gcd ignored eos files with fid>4294967295
@@ -81,7 +99,7 @@ CTA v3.1-14.
 
 ## Summary
 
-This version is a transition version between CTA v3.1-14 and CTA v4.0. The functionalities of the v4.0 are implemented, 
+This version is a transition version between CTA v3.1-14 and CTA v4.0. The functionalities of the v4.0 are implemented,
 but at the database level, the columns to be deleted have been put as NULLABLE and the NOT NULL columns to be added as NULLABLE.
 The xrootd-ssi-protobuf-interface is not up-to-date with the CTA v4.0: deprecated fields have not been removed.
 
@@ -190,7 +208,7 @@ This release contains the CTA software Recommended Access Order (RAO) implemente
 - Upgraded EOS to 4.8.24-1
 - Upgraded xrootd to 4.12.5-1
 - cta-admin repack ls tabular output improvements
-- Repack management execution can be disabled via the cta-taped configuration file 
+- Repack management execution can be disabled via the cta-taped configuration file
 - cta/CTA#907 Maintenance process can be disabled via the cta-taped configuration file
 
 # Modifications
@@ -200,7 +218,7 @@ This release contains the CTA software Recommended Access Order (RAO) implemente
 ### Bug fixes
 
 - cta/CTA#901 cta-admin tapefile ls too slow
-- cta/CTA#895 [catalogue] RdbmsCatalogue::deleteLogicalLibrary does not delete empty logical library 
+- cta/CTA#895 [catalogue] RdbmsCatalogue::deleteLogicalLibrary does not delete empty logical library
 - utils::trimString() now returns an empty string if the string passed in parameter contains only white-space characters
 - Repack request and sub-requests are now unowned from their Agent when completed
 
@@ -413,7 +431,7 @@ command-line tool based on operator requests and removes the deprecated
 
 ## Summary
 
-### Features 
+### Features
 
 - New schema version 2.0
   - *VIRTUAL_ORGANIZATION* has its own table
@@ -427,7 +445,7 @@ command-line tool based on operator requests and removes the deprecated
 
 ## Summary
 
-### Features 
+### Features
 
 - cta-admin tapefile ls command: list the tape files located in a specific tape
 - New schema verification tool (cta-catalogue-schema-verify)
@@ -450,7 +468,7 @@ command-line tool based on operator requests and removes the deprecated
 
 - The scheduler does not return a mount if a tape is disabled (unless if the tape is repacked with the --disabledtape flag set)
 
-### Improvements 
+### Improvements
 
 - CASTOR-To-CTA migration improvements
 - Better cta-admin parameter checking and column formatting